Fixed warnings, small bugs
Caching broke again?!?

+++ src/apis/config.cpp
@@ -88,13 +88,13 @@ static int config_get(lua_State *L) {
         else lua_pushnil(L);
     } else if (strcmp(name, "http_whitelist") == 0) {
         lua_createtable(L, config.http_whitelist.size(), 0);
-        for (int i = 0; i < config.http_whitelist.size(); i++) {
+        for (size_t i = 0; i < config.http_whitelist.size(); i++) {
             lua_pushstring(L, config.http_whitelist[i].c_str());
             lua_rawseti(L, -2, i+1);
         }
     } else if (strcmp(name, "http_blacklist") == 0) {
         lua_createtable(L, config.http_blacklist.size(), 0);
-        for (int i = 0; i < config.http_blacklist.size(); i++) {
+        for (size_t i = 0; i < config.http_blacklist.size(); i++) {
             lua_pushstring(L, config.http_blacklist[i].c_str());
             lua_rawseti(L, -2, i+1);
         }
